lxpanel: fix gtk3 geometry issues
When the wnck pager is displayed the panel's width was set to the sum of
all widgets, not the screen width.
The with of a single task in the task bar was not reduced when all task's
did no longer fit in the allocated taskbar width resulting in a panel which
extended past the right side of the screen.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>